#1183 using NERFormModal
