+++ src/frontend/src/apis/teams.api.ts
@@ -38,7 +38,7 @@ export const setTeamHead = (id: string, userId: number) => {
 };
 
 export const deleteTeam = (id: string) => {
-  return axios.post(apiUrls.teamsDelete(id));
+  return axios.post<{ message: string }>(apiUrls.teamsDelete(id));
 };
 
 export const setTeamLeads = (id: string, userIds: number[]) => {
