+++ src/frontend/src/pages/TeamsPage/Teams.tsx
@@ -6,12 +6,12 @@
 import { Route, Switch } from 'react-router-dom';
 import { routes } from '../../utils/routes';
 import TeamsPage from './TeamsPage';
-import TeamDetailPage from './TeamSpecificPage';
+import TeamSpecificPage from './TeamSpecificPage';
 
 const Teams: React.FC = () => {
   return (
     <Switch>
-      <Route path={routes.TEAMS_BY_ID} component={TeamDetailPage} />
+      <Route path={routes.TEAMS_BY_ID} component={TeamSpecificPage} />
       <Route path={routes.TEAMS} component={TeamsPage} />
     </Switch>
   );
